// types/primarie.ts
export interface Primarie {
  nrcrt: number
  numePrimarie: string
  populatie: number
  Adresa: string
  demail: string
  telefonprimariegeneral: string
  nrfax: number
  Primar: string
  telefonprimar: number
  Viceprimar: string
  telefonviceprimar: number
  Viceprimar2?: string
  telefonviceprimar2?: number
  "Şef SVSU": string
  telefonsvsu: number
  coordinates: {
    latitude: number
    longitude: number
  }
}